+++ scheme_port.c
@@ -142,7 +142,7 @@ scheme_make_input_port (Scheme_Object *subtype,
 			void *data,
 			int (*getc_fun) (Scheme_Input_Port*),
 			void (*ungetc_fun) (int, Scheme_Input_Port*),
-			long (*char_ready_fun) (Scheme_Input_Port*),
+			size_t (*char_ready_fun) (Scheme_Input_Port*),
 			void (*close_fun) (Scheme_Input_Port*))
 {
   Scheme_Input_Port *ip;
@@ -232,7 +232,7 @@ file_ungetc (int ch, Scheme_Input_Port *port)
   ungetc (ch, (FILE *)port->port_data);
 }
 
-static long
+static size_t
 file_char_ready (Scheme_Input_Port *port)
 {
   FILE *fp = (FILE *) port->port_data;
@@ -242,7 +242,7 @@ file_char_ready (Scheme_Input_Port *port)
   return (fp->_egptr - fp->_gptr);
 #else
   scheme_warning ("char-ready? always returns #f on this platform");
-  return ((long)scheme_false);
+  return ((size_t)scheme_false);
 #endif
 }
 
@@ -301,7 +301,7 @@ string_ungetc (int ch, Scheme_Input_Port *port)
     }
 }
 
-static long
+static size_t
 string_char_ready (Scheme_Input_Port *port)
 {
   Scheme_Indexed_String *is;
